requirements.txt
aniso8601==4.0.1
azf-wsgi==0.3.0
azure-functions==1.0.8
Click==7.0
Flask==0.10.1
Flask-Cors==3.0.7
Flask-HTTPAuth==2.3.0
Flask-RESTful==0.3.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
numpy==1.15.3
pandas==0.23.4
psycopg2==2.7.5
python-dateutil==2.7.5
pytz==2018.6
six==1.9.0
SQLAlchemy==1.2.13
Werkzeug==0.9.6
